Rework sidebar and quest selection
You can now select quest helpers from the sidebar. This will update to show only available helpers.

You can stop a helper from a helper's sidebar now.

Restructured more complex isComplete checks to more accurately reflect if a helper should be considered 'complete' or not.
