+++ src/main/java/com/questhelper/QuestHelperQuest.java
@@ -27,7 +27,7 @@ public enum QuestHelperQuest
 	SHEEP_SHEARER(315, "Sheep Shearer", QuestVarPlayer.QUEST_SHEEP_SHEARER),
 	SHIELD_OF_ARRAV_PHOENIX_GANG(316, "Shield of Arrav - Phoenix Gang", QuestVarPlayer.QUEST_SHIELD_OF_ARRAV),
 	SHIELD_OF_ARRAV_BLACK_ARM_GANG(316, "Shield of Arrav - Black Arm Gang", QuestVarPlayer.QUEST_SHIELD_OF_ARRAV_STATE_146),
-	VAMPYRE_SLAYER(317, "Vampyre Slayer", QuestVarPlayer.QUEST_VAMPIRE_SLAYER),
+	VAMPYRE_SLAYER(1278, "Vampyre Slayer", QuestVarPlayer.QUEST_VAMPYRE_SLAYER),
 	WITCHS_POTION(318, "Witch's Potion", QuestVarPlayer.QUEST_WITCHS_POTION),
 	X_MARKS_THE_SPOT(550, "X Marks the Spot", QuestVarbits.QUEST_X_MARKS_THE_SPOT),
 
@@ -116,14 +116,14 @@ public enum QuestHelperQuest
 	RATCATCHERS(412, "Ratcatchers", QuestVarbits.QUEST_RATCATCHERS),
 	RECIPE_FOR_DISASTER(413, "Recipe for Disaster", QuestVarbits.QUEST_RECIPE_FOR_DISASTER),
 	RECIPE_FOR_DISASTER_START(413, "RFD - Start", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_START),
-//	RECIPE_FOR_DISASTER_DWARF(413, "Recipe for Disaster - Dwarf", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_DWARF),
-//	RECIPE_FOR_DISASTER_WARTFACE_AND_BENTNOZE(413, "Recipe for Disaster - Wartface & Bentnoze", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_WARTFACE_AND_BENTNOZE),
+//	RECIPE_FOR_DISASTER_DWARF(413, "RFD - Dwarf", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_DWARF),
+//	RECIPE_FOR_DISASTER_WARTFACE_AND_BENTNOZE(413, "RFD - Wartface & Bentnoze", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_WARTFACE_AND_BENTNOZE),
 	RECIPE_FOR_DISASTER_PIRATE_PETE(413, "RFD - Pirate Pete", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_PIRATE_PETE),
 	RECIPE_FOR_DISASTER_LUMBRIDGE_GUIDE(413, "RFD - Lumbridge Guide", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_LUMBRIDGE_GUIDE),
-//	RECIPE_FOR_DISASTER_EVIL_DAVE(413, "Recipe for Disaster - Evil Dave", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_EVIL_DAVE),
-//	RECIPE_FOR_DISASTER_MONKEY_AMBASSADOR(413, "Recipe for Disaster - Monkey Ambassador", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_MONKEY_AMBASSADOR),
-//	RECIPE_FOR_DISASTER_SIR_AMIK_VARZE(413, "Recipe for Disaster - Sir Amik Varze", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_SIR_AMIK_VARZE),
-//	RECIPE_FOR_DISASTER_UGLOGWEE(413, "Recipe for Disaster - Skrach Uglogwee", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_UGLOGWEE),
+//	RECIPE_FOR_DISASTER_EVIL_DAVE(413, "RFD - Evil Dave", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_EVIL_DAVE),
+//	RECIPE_FOR_DISASTER_MONKEY_AMBASSADOR(413, "RFD - Monkey Ambassador", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_MONKEY_AMBASSADOR),
+//	RECIPE_FOR_DISASTER_SIR_AMIK_VARZE(413, "RFD - Sir Amik Varze", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_SIR_AMIK_VARZE),
+//	RECIPE_FOR_DISASTER_UGLOGWEE(413, "RFD - Skrach Uglogwee", QuestVarbits.QUEST_RECIPE_FOR_DISASTER_UGLOGWEE),
 	RECRUITMENT_DRIVE(414, "Recruitment Drive", QuestVarbits.QUEST_RECRUITMENT_DRIVE),
 	REGICIDE(415, "Regicide", QuestVarPlayer.QUEST_REGICIDE),
 	ROVING_ELVES(416, "Roving Elves", QuestVarPlayer.QUEST_ROVING_ELVES),
