+++ src/main/java/com/questhelper/QuestVarPlayer.java
@@ -32,7 +32,7 @@ public enum QuestVarPlayer
 	QUEST_SHEEP_SHEARER(179),
 	QUEST_SHIELD_OF_ARRAV(145),
 	QUEST_SHIELD_OF_ARRAV_STATE_146(146),
-	QUEST_VAMPIRE_SLAYER(178),
+	QUEST_VAMPYRE_SLAYER(178),
 	QUEST_WITCHS_POTION(67),
 
 	/**
